Calling test clears out analysis table. Make sure to call testfunction every time you access analysis

+++ select.sh
@@ -4,7 +4,9 @@
 timeout 10s mysql --host=dolphin --user=csci375team6 --password=3jni3edn << EOF
 use csci375team6_povCal;
 
-delete from analysis;
+
+select * from analysis;
+--delete from analysis where ownerId = '1';
